coio/say: fix getaddrinfo error handling on macOS
Before this patch, branch when getaddrinfo() returns error codes
couldn't be reached on macOS, because they are greater than 0 on
macOS (assumption "rc < 0" in commit ea1da04 is incorrect for
macOS).

Note: diag_log() in say.c was added, because otherwise it will be
hid in the case of panic(). Also, two diag_set() in
syslog_connect_unix() was added to avoid asserts in this
diag_log().

Needed for #4138